refactor(componente-tabela): parametrizando título do botão de cadastro, e movendo GenericContext para pasta contexts
